Interaction-layer dead-code sweep + small dedups (~190 LOC, issue #231)
Dead code removed:
- `compareLayers` from interaction-priority.ts (private, zero callers)
- `isDragId` from drag-ids.ts (private, zero callers)
- `isCommandShortcutKey` from gesture-utils.ts (private, zero callers)
- `update()` export from interaction-controller.ts (exported but uncalled)
- `entityResizesAutomatically` always-false stub from hit-test.ts; inlined
  the call site in collectResizeHandles to a simple skip (never-true guard)
- Unused props on SelectionResizeGrid (id, width, height, canvasX, canvasY,
  zoom, minWidth, minHeight, onResize, aspectRatioResizeMode); trimmed all
  call sites in SelectionOutlineLayer and removed the now-dead computed vars
  (zoom, minWidth, minHeight, aspectRatioResizeMode) and their imports
- `void resolveCanvasColor` speculative import in GroupRenameLabel.tsx

Small dedups landed:
- (a) Shared `canvasInteractionModeKind(state)` in gesture-utils.ts replaces
  the identical switch tables in `interactionModeKey` (focus-reconciler-runtime)
  and `predicateInteractionKind` (selection-controller)
- (b) `fileDisplayName(file)` in entityConstants.ts replaces the duplicated
  basename + suffix-strip in FileChrome.tsx and FilePopup.tsx
- (c) `layoutToHitInputs(layout)` helper in useCanvasPointerRouter.ts
  replaces the two identical HitInputs literal constructions
- (d) `DRAG_THRESHOLD = 4` exported from gesture-utils.ts replaces four
  module-local constants in useCanvasPointerRouter.ts plus the local copies
  in GroupRenameLabel.tsx (GROUP_DRAG_THRESHOLD) and App.tsx
  (COMMENT_DRAG_THRESHOLD)

Closes #231
