+++ src/renderer/above-view/FilePopup.tsx
@@ -11,16 +11,9 @@ import type {
 import { CanvasItemPopup } from './CanvasItemPopup'
 import { DistributeAction } from './DistributeAction'
 import { InlineEditLabel } from '../shared/InlineEditLabel'
-import { MARKDOWN_EXTENSIONS, WIREFRAME_EXTENSIONS } from '../canvas-bg/entityConstants'
+import { fileDisplayName } from '../canvas-bg/entityConstants'
 import { renderPopupContributions } from './file-popup-contributions'
 import { POPUP_OFFSET_Y, usePopupDelayedKey } from './usePopupDelayedKey'
-
-function displayNameFor(file: string): string {
-  const base = file.split('/').pop() ?? file
-  if (WIREFRAME_EXTENSIONS.test(file)) return base.replace(/\.wireframe\.json$/i, '')
-  if (MARKDOWN_EXTENSIONS.test(file)) return base.replace(/\.md$/i, '')
-  return base
-}
 
 export function FilePopup({
   api,
@@ -92,7 +85,7 @@ export function FilePopup({
                   ) : null}
                   <CanvasItemPopup.Section grow>
                     <InlineEditLabel
-                      value={displayNameFor(single.file)}
+                      value={fileDisplayName(single.file)}
                       isEditing={isRenaming}
                       onStartEdit={() => setIsRenaming(true)}
                       onCommit={(next) => {
