+++ src/renderer/canvas-bg/entityConstants.ts
@@ -22,8 +22,15 @@ export {
   WIREFRAME_EXTENSIONS,
   HTML_EXTENSIONS,
 } from '../../shared/file-extensions'
-import { IMAGE_EXTENSIONS, VIDEO_EXTENSIONS } from '../../shared/file-extensions'
+import { IMAGE_EXTENSIONS, VIDEO_EXTENSIONS, MARKDOWN_EXTENSIONS, WIREFRAME_EXTENSIONS } from '../../shared/file-extensions'
 import { RESIZE_HANDLE_VISUAL_PX } from '../../shared/canvas-hit-geometry'
+
+export function fileDisplayName(file: string): string {
+  const base = file.split('/').pop() ?? file
+  if (WIREFRAME_EXTENSIONS.test(file)) return base.replace(/\.wireframe\.json$/i, '')
+  if (MARKDOWN_EXTENSIONS.test(file)) return base.replace(/\.md$/i, '')
+  return base
+}
 
 /** Images/videos: lock aspect unless Shift. Other files: free resize unless Shift (then lock). */
 export function aspectRatioResizeModeForCanvasFile(filePath: string): AspectRatioResizeMode {
